requirements.txt
azure-functions
azure-search-documents==11.4.0
azure-core
python-dotenv